fantasy-football-api/pyproject.toml
[project]
name = "fantasy-football-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "espn-api>=0.45.1",
    "fastapi>=0.119.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
